fix(media): bound CLI reference image input
